Safari-609.1.11.1/LayoutTests/fast/animation/request-animation-frame-time-unit-expected.txt

Tests time unit of requestAnimationFrame callback

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS isTimeUnitInMs is true
PASS successfullyParsed is true

TEST COMPLETE
